+++ library/inc/controls/class-WM_Customizer_Hidden.php
@@ -1,14 +1,15 @@
 <?php
 /**
- * Skinning System
+ * Theme Customizer Input Fields
  *
  * Customizer hidden input field.
  *
  * @package     WebMan WordPress Theme Framework
- * @subpackage  Skinning System
+ * @subpackage  Theme Customizer
  * @copyright   2014 WebMan - Oliver Juhas
  *
- * @since       3.1
+ * @since    3.1
+ * @version  4.0
  */
 
 
@@ -22,9 +23,9 @@ class WM_Customizer_Hidden extends WP_Customize_Control {
 
 	public function render_content() {
 		?>
-		<textarea <?php $this->link(); ?>>
-			<?php echo esc_textarea( $this->value() ); ?>
-		</textarea>
+
+		<textarea <?php $this->link(); ?>><?php echo esc_textarea( $this->value() ); ?></textarea>
+
 		<?php
 	}