+++ library/inc/controls/class-WM_Customizer_Image.php
@@ -1,14 +1,15 @@
 <?php
 /**
- * Skinning System
+ * Theme Customizer Input Fields
  *
  * Customizer image insert.
  *
  * @package     WebMan WordPress Theme Framework
- * @subpackage  Skinning System
+ * @subpackage  Theme Customizer
  * @copyright   2014 WebMan - Oliver Juhas
  *
- * @since       3.1
+ * @since    3.1
+ * @version  4.0
  */
 
 
@@ -16,7 +17,7 @@
 /**
  * Add uploaded images tab to Image control
  */
-class WM_Customize_Image extends WP_Customize_Image_Control {
+class WM_Customizer_Image extends WP_Customize_Image_Control {
 
 	/**
 	 * Adding an .ico into supported image file formats
@@ -52,6 +53,6 @@ class WM_Customize_Image extends WP_Customize_Image_Control {
 		}
 	}
 
-} // /WM_Customize_Image
+} // /WM_Customizer_Image
 
 ?>