+++ library/inc/controls/class-WM_Customizer_Textarea.php
@@ -1,14 +1,15 @@
 <?php
 /**
- * Skinning System
+ * Theme Customizer Input Fields
  *
  * Customizer textarea field.
  *
  * @package     WebMan WordPress Theme Framework
- * @subpackage  Skinning System
+ * @subpackage  Theme Customizer
  * @copyright   2014 WebMan - Oliver Juhas
  *
- * @since       3.1
+ * @since    3.1
+ * @version  4.0
  */
 
 
@@ -22,12 +23,13 @@ class WM_Customizer_Textarea extends WP_Customize_Control {
 
 	public function render_content() {
 		?>
+
 		<label>
 			<span class="customize-control-title"><?php echo $this->label; ?></span>
-			<textarea cols="20" rows="4" <?php $this->link(); ?>>
-				<?php echo esc_textarea( $this->value() ); ?>
-			</textarea>
+			<?php if ( $this->description ) : ?><span class="description customize-control-description"><?php echo $this->description; ?></span><?php endif; ?>
+			<textarea name="<?php echo $this->id; ?>" cols="20" rows="4" <?php $this->link(); ?>><?php echo esc_textarea( $this->value() ); ?></textarea>
 		</label>
+
 		<?php
 	}